+++ packages/eyes-protractor/src/ProtractorWrappedElement.js
@@ -1,0 +1,6 @@
+const {EyesWrappedElement} = require('@applitools/eyes-sdk-core')
+const SpecWrappedElement = require('./SpecWrappedElement')
+
+const ProtractorWrappedElement = EyesWrappedElement.specialize(SpecWrappedElement)
+
+module.exports = ProtractorWrappedElement