+++ packages/eyes-sdk-core/lib/EyesJsSnippets.js
@@ -36,8 +36,8 @@ const GET_ELEMENT_RECT = `
   var computedStyle = window.getComputedStyle(element);
   var isFixed = isFixedElement(element);
   return {
-    x: isFixed ? element.offsetLeft : rect.left + (window.scrollX || window.pageXOffset),
-    y: isFixed ? element.offsetTop : rect.top + (window.scrollY || window.pageYOffset),
+    x: isFixed ? rect.left : rect.left + (window.scrollX || window.pageXOffset),
+    y: isFixed ? rect.top : rect.top + (window.scrollY || window.pageYOffset),
     width: rect.width,
     height: rect.height
   };
@@ -233,7 +233,7 @@ const GET_CURRENT_CONTEXT_INFO = `
   return {
     isRoot: isRoot,
     isCORS: isCORS,
-    document: document.documentElement,
+    contentDocument: document.querySelector('html'),
     selector: selector,
   };
 `