+++ packages/eyes-sdk-core/lib/config/BrowserType.js
@@ -20,6 +20,7 @@ const BrowserType = {
   SAFARI_ONE_VERSION_BACK: 'safari-one-version-back',
   SAFARI_TWO_VERSIONS_BACK: 'safari-two-versions-back',
   EDGE_CHROMIUM_ONE_VERSION_BACK: 'edgechromium-one-version-back',
+  EDGE_CHROMIUM_TWO_VERSIONS_BACK: 'edgechromium-two-versions-back',
 }
 
 Object.freeze(BrowserType)