+++ packages/eyes-images/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## Unreleased
+- updated to @applitools/eyes-sdk-core@11.0.2 (from v10.2.0)
+- updated to @applitools/eyes-sdk-core@11.0.4 (from v11.0.2)