+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 11.0.4 - 2020/6/4
+
+- fix marking the scroll root element
+- support correct driver API in DomCapture
+
+## 11.0.3 - 2020/6/3
+
+- added EDGE_CHROMIUM_TWO_VERSIONS_BACK to BrowserType
+- updated to @applitools/dom-capture@7.2.4 (from v7.2.0)
+
 ## 11.0.2 - 2020/6/2
 
 - avoid 404 error in internet explorer due to parallel script executions